Moved class for converting uima annotations into library-internal annotation classes into its own package
